+++ dozer/cogs/general.py
@@ -122,6 +122,30 @@ class General(Cog):
 		"""Allows a member to change their nickname."""
 		await discord.Member.edit(ctx.author, nick=nicktochangeto)
 		await ctx.send("Nick successfully changed to " + nicktochangeto)
+	@command()
+	async def tba (self,ctx,task,teamnum):
+		"""Pulls Team Data from TBA. Subcommand Team: Pulls team data."""
+		parser = tbapi.TBAParser('0000', 'Dozer', 'Beta 0.7')
+		teamdata = parser.get_team('frc' + teamnum)
+		if task == 'team':
+			guild = ctx.guild
+			e = discord.Embed(color=blurple)
+			e.add_field(name='Team Name', value=teamdata.nickname)
+			e.add_field(name='Sponsors', value=teamdata.name)
+			e.add_field(name='Team Number', value=teamdata.number)
+			e.add_field(name='Team Key', value=teamdata.key)
+			e.add_field(name='Team Location', value=teamdata.location)
+			e.add_field(name='Rookie Year', value=teamdata.rookie_year)
+			e.add_field(name='Team Motto', value=teamdata.motto)
+			e.add_field(name='Team Website', value=teamdata.website)
+			e.add_field(name='TBA Page', value='https://www.thebluealliance.com/team/' + teamnum)
+			await ctx.send(embed=e)
+		if task == 'raw':
+			await ctx.send(teamdata.raw)
+	@command()
+	async def rip(self,ctx):
+		"""Rest in Piece"""
+		await ctx.send('RIP http://worldartsme.com/images/tombstone-clipart-1.jpg')
 
 def setup(bot):
 	bot.remove_command('help')
